add grpc proxying to enable bring-your-own-proto
