+++ pkg/diagnostics/grpc_tracing.go
@@ -10,6 +10,8 @@ import (
 	"fmt"
 	"strings"
 
+	grpc_middleware "github.com/grpc-ecosystem/go-grpc-middleware"
+	"github.com/pkg/errors"
 	"go.opencensus.io/trace"
 	"go.opencensus.io/trace/propagation"
 	"google.golang.org/grpc"
@@ -23,7 +25,10 @@ import (
 	runtimev1pb "github.com/dapr/dapr/pkg/proto/runtime/v1"
 )
 
-const grpcTraceContextKey = "grpc-trace-bin"
+const (
+	grpcTraceContextKey = "grpc-trace-bin"
+	GRPCProxyAppIDKey   = "dapr-app-id"
+)
 
 // GRPCTraceUnaryServerInterceptor sets the trace context or starts the trace client span based on request
 func GRPCTraceUnaryServerInterceptor(appID string, spec config.TracingSpec) grpc.UnaryServerInterceptor {
@@ -47,26 +52,9 @@ func GRPCTraceUnaryServerInterceptor(appID string, spec config.TracingSpec) grpc
 		}
 
 		ctx, span = trace.StartSpanWithRemoteParent(ctx, spanName, sc, sampler, spanKind)
-
-		var prefixedMetadata map[string]string
-		if span.SpanContext().TraceOptions.IsSampled() {
-			// users can add dapr- prefix if they want to see the header values in span attributes.
-			prefixedMetadata = userDefinedMetadata(ctx)
-		}
-
 		resp, err := handler(ctx, req)
 
-		if span.SpanContext().TraceOptions.IsSampled() {
-			// Populates dapr- prefixed header first
-			AddAttributesToSpan(span, prefixedMetadata)
-			spanAttr := spanAttributesMapFromGRPC(appID, req, info.FullMethod)
-			AddAttributesToSpan(span, spanAttr)
-
-			// Correct the span name based on API.
-			if sname, ok := spanAttr[daprAPISpanNameInternal]; ok {
-				span.SetName(sname)
-			}
-		}
+		addSpanMetadataAndUpdateStatus(ctx, span, info.FullMethod, appID, req, false)
 
 		// Add grpc-trace-bin header for all non-invocation api's
 		if info.FullMethod != "/dapr.proto.runtime.v1.Dapr/InvokeService" {
@@ -78,6 +66,72 @@ func GRPCTraceUnaryServerInterceptor(appID string, spec config.TracingSpec) grpc
 		span.End()
 
 		return resp, err
+	}
+}
+
+// GRPCTraceStreamServerInterceptor sets the trace context or starts the trace client span based on request
+func GRPCTraceStreamServerInterceptor(appID string, spec config.TracingSpec) grpc.StreamServerInterceptor {
+	return func(srv interface{}, ss grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
+		var span *trace.Span
+		spanName := info.FullMethod
+
+		ctx := ss.Context()
+		md, _ := metadata.FromIncomingContext(ctx)
+
+		vals := md.Get(GRPCProxyAppIDKey)
+		if len(vals) == 0 {
+			return errors.Errorf("cannot proxy request: missing %s metadata", GRPCProxyAppIDKey)
+		}
+
+		targetID := vals[0]
+		wrapped := grpc_middleware.WrapServerStream(ss)
+		sc, _ := SpanContextFromIncomingGRPCMetadata(ctx)
+		sampler := diag_utils.TraceSampler(spec.SamplingRate)
+
+		var spanKind trace.StartOption
+
+		if appID == targetID {
+			spanKind = trace.WithSpanKind(trace.SpanKindServer)
+		} else {
+			spanKind = trace.WithSpanKind(trace.SpanKindClient)
+		}
+
+		ctx, span = trace.StartSpanWithRemoteParent(ctx, spanName, sc, sampler, spanKind)
+		wrapped.WrappedContext = ctx
+		err := handler(srv, wrapped)
+
+		addSpanMetadataAndUpdateStatus(ctx, span, info.FullMethod, appID, nil, true)
+
+		UpdateSpanStatusFromGRPCError(span, err)
+		span.End()
+
+		return err
+	}
+}
+
+func addSpanMetadataAndUpdateStatus(ctx context.Context, span *trace.Span, fullMethod, appID string, req interface{}, stream bool) {
+	var prefixedMetadata map[string]string
+	if span.SpanContext().TraceOptions.IsSampled() {
+		// users can add dapr- prefix if they want to see the header values in span attributes.
+		prefixedMetadata = userDefinedMetadata(ctx)
+	}
+
+	if span.SpanContext().TraceOptions.IsSampled() {
+		// Populates dapr- prefixed header first
+		AddAttributesToSpan(span, prefixedMetadata)
+
+		spanAttr := map[string]string{}
+		if !stream {
+			spanAttr = spanAttributesMapFromGRPC(appID, req, fullMethod)
+			AddAttributesToSpan(span, spanAttr)
+		} else {
+			spanAttr[daprAPISpanNameInternal] = fullMethod
+		}
+
+		// Correct the span name based on API.
+		if sname, ok := spanAttr[daprAPISpanNameInternal]; ok {
+			span.SetName(sname)
+		}
 	}
 }
 
@@ -145,6 +199,7 @@ func SpanContextToGRPCMetadata(ctx context.Context, spanContext trace.SpanContex
 	if (trace.SpanContext{}) == spanContext {
 		return ctx
 	}
+
 	traceContextBinary := propagation.Binary(spanContext)
 	return metadata.AppendToOutgoingContext(ctx, grpcTraceContextKey, string(traceContextBinary))
 }
