+++ pkg/grpc/grpc.go
@@ -55,7 +55,7 @@ func (g *Manager) SetAuthenticator(auth security.Authenticator) {
 
 // CreateLocalChannel creates a new gRPC AppChannel
 func (g *Manager) CreateLocalChannel(port, maxConcurrency int, spec config.TracingSpec, sslEnabled bool) (channel.AppChannel, error) {
-	conn, err := g.GetGRPCConnection(fmt.Sprintf("127.0.0.1:%v", port), "", "", true, false, sslEnabled)
+	conn, err := g.GetGRPCConnection(context.TODO(), fmt.Sprintf("127.0.0.1:%v", port), "", "", true, false, sslEnabled)
 	if err != nil {
 		return nil, errors.Errorf("error establishing connection to app grpc on port %v: %s", port, err)
 	}
@@ -66,7 +66,7 @@ func (g *Manager) CreateLocalChannel(port, maxConcurrency int, spec config.Traci
 }
 
 // GetGRPCConnection returns a new grpc connection for a given address and inits one if doesn't exist
-func (g *Manager) GetGRPCConnection(address, id string, namespace string, skipTLS, recreateIfExists, sslEnabled bool) (*grpc.ClientConn, error) {
+func (g *Manager) GetGRPCConnection(ctx context.Context, address, id string, namespace string, skipTLS, recreateIfExists, sslEnabled bool, customOpts ...grpc.DialOption) (*grpc.ClientConn, error) {
 	g.lock.RLock()
 	if val, ok := g.connectionPool[address]; ok && !recreateIfExists {
 		g.lock.RUnlock()
@@ -112,7 +112,7 @@ func (g *Manager) GetGRPCConnection(address, id string, namespace string, skipTL
 		transportCredentialsAdded = true
 	}
 
-	ctx, cancel := context.WithTimeout(context.Background(), dialTimeout)
+	ctx, cancel := context.WithTimeout(ctx, dialTimeout)
 	defer cancel()
 
 	dialPrefix := GetDialAddressPrefix(g.mode)
@@ -128,6 +128,7 @@ func (g *Manager) GetGRPCConnection(address, id string, namespace string, skipTL
 		opts = append(opts, grpc.WithInsecure())
 	}
 
+	opts = append(opts, customOpts...)
 	conn, err := grpc.DialContext(ctx, dialPrefix+address, opts...)
 	if err != nil {
 		return nil, err
