+++ pkg/grpc/grpc_test.go
@@ -6,6 +6,7 @@
 package grpc
 
 import (
+	"context"
 	"crypto/x509"
 	"fmt"
 	"testing"
@@ -50,9 +51,10 @@ func TestGetGRPCConnection(t *testing.T) {
 		assert.NotNil(t, m)
 		port := 55555
 		sslEnabled := false
-		conn, err := m.GetGRPCConnection(fmt.Sprintf("127.0.0.1:%v", port), "", "", true, true, sslEnabled)
+		ctx := context.TODO()
+		conn, err := m.GetGRPCConnection(ctx, fmt.Sprintf("127.0.0.1:%v", port), "", "", true, true, sslEnabled)
 		assert.NoError(t, err)
-		conn2, err2 := m.GetGRPCConnection(fmt.Sprintf("127.0.0.1:%v", port), "", "", true, true, sslEnabled)
+		conn2, err2 := m.GetGRPCConnection(ctx, fmt.Sprintf("127.0.0.1:%v", port), "", "", true, true, sslEnabled)
 		assert.NoError(t, err2)
 		assert.Equal(t, connectivity.Shutdown, conn.GetState())
 		conn2.Close()
@@ -63,7 +65,8 @@ func TestGetGRPCConnection(t *testing.T) {
 		assert.NotNil(t, m)
 		port := 55555
 		sslEnabled := true
-		_, err := m.GetGRPCConnection(fmt.Sprintf("127.0.0.1:%v", port), "", "", true, true, sslEnabled)
+		ctx := context.TODO()
+		_, err := m.GetGRPCConnection(ctx, fmt.Sprintf("127.0.0.1:%v", port), "", "", true, true, sslEnabled)
 		assert.NoError(t, err)
 	})
 }
